feat: initial commit
First commit for the Rust for Lunch website, lunch.rs.

This contains a (very) basic web-site with a first meetup date listed.
The web-site is built with the Zola static site generator.

As well as a main page, there is an about page that details the code of
contact and contact details.